+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/Synchronizer.kt
@@ -15,8 +15,10 @@ import cash.z.ecc.android.sdk.internal.model.ext.toBlockHeight
 import cash.z.ecc.android.sdk.model.Account
 import cash.z.ecc.android.sdk.model.BlockHeight
 import cash.z.ecc.android.sdk.model.PercentDecimal
+import cash.z.ecc.android.sdk.model.Proposal
 import cash.z.ecc.android.sdk.model.TransactionOverview
 import cash.z.ecc.android.sdk.model.TransactionRecipient
+import cash.z.ecc.android.sdk.model.TransactionSubmitResult
 import cash.z.ecc.android.sdk.model.UnifiedSpendingKey
 import cash.z.ecc.android.sdk.model.WalletBalance
 import cash.z.ecc.android.sdk.model.Zatoshi
@@ -168,6 +170,64 @@ interface Synchronizer {
     suspend fun getTransparentAddress(account: Account): String
 
     /**
+     * Creates a proposal for transferring funds to the given recipient.
+     *
+     * @param account the account from which to transfer funds.
+     * @param recipient the recipient's address.
+     * @param amount the amount of zatoshi to send.
+     * @param memo the optional memo to include as part of the proposal's transactions.
+     *
+     * @return the proposal or an exception
+     */
+    suspend fun proposeTransfer(
+        account: Account,
+        recipient: String,
+        amount: Zatoshi,
+        memo: String = ""
+    ): Proposal
+
+    /**
+     * Creates a proposal for shielding any transparent funds received by the given account.
+     *
+     * @param account the account for which to shield funds.
+     * @param shieldingThreshold the minimum transparent balance required before a
+     *                           proposal will be created.
+     * @param memo the optional memo to include as part of the proposal's transactions.
+     * @param transparentReceiver a specific transparent receiver within the account that
+     *                            should be the source of transparent funds. Default is
+     *                            null which will select whichever of the account's
+     *                            transparent receivers has funds to shield.
+     *
+     * @return the proposal, or null if the transparent balance that would be shielded is
+     *         zero or below `shieldingThreshold`.
+     *
+     * @throws Exception if `transparentReceiver` is null and there are transparent funds
+     *         in more than one of the account's transparent receivers.
+     */
+    suspend fun proposeShielding(
+        account: Account,
+        shieldingThreshold: Zatoshi,
+        memo: String = ZcashSdk.DEFAULT_SHIELD_FUNDS_MEMO_PREFIX,
+        transparentReceiver: String? = null
+    ): Proposal?
+
+    /**
+     * Creates the transactions in the given proposal.
+     *
+     * @param proposal the proposal for which to create transactions.
+     * @param usk the unified spending key associated with the account for which the
+     *            proposal was created.
+     *
+     * @return a flow of result objects for the transactions that were created as part of
+     *         the proposal, indicating whether they were submitted to the network or if
+     *         an error occurred.
+     */
+    suspend fun createProposedTransactions(
+        proposal: Proposal,
+        usk: UnifiedSpendingKey
+    ): Flow<TransactionSubmitResult>
+
+    /**
      * Sends zatoshi.
      *
      * @param usk the unified spending key associated with the notes that will be spent.
@@ -180,6 +240,13 @@ interface Synchronizer {
      * useful for updating the UI without needing to poll. Of course, polling is always an option
      * for any wallet that wants to ignore this return value.
      */
+    @Deprecated(
+        message = "Upcoming SDK 2.1 will create multiple transactions at once for some recipients.",
+        replaceWith =
+            ReplaceWith(
+                "createProposedTransactions(proposeTransfer(usk.account, toAddress, amount, memo), usk)"
+            )
+    )
     suspend fun sendToAddress(
         usk: UnifiedSpendingKey,
         amount: Zatoshi,
@@ -187,6 +254,13 @@ interface Synchronizer {
         memo: String = ""
     ): Long
 
+    @Deprecated(
+        message = "Upcoming SDK 2.1 will create multiple transactions at once for some recipients.",
+        replaceWith =
+            ReplaceWith(
+                "proposeShielding(usk.account, shieldingThreshold, memo)?.let { createProposedTransactions(it, usk) }"
+            )
+    )
     suspend fun shieldFunds(
         usk: UnifiedSpendingKey,
         memo: String = ZcashSdk.DEFAULT_SHIELD_FUNDS_MEMO_PREFIX