+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/model/Proposal.kt
@@ -20,7 +20,7 @@ class Proposal(
 
             // Check for type errors eagerly, to ensure that the caller won't
             // encounter these errors later.
-            typed.feeRequired()
+            typed.totalFeeRequired()
 
             return typed
         }
@@ -34,9 +34,21 @@ class Proposal(
     }
 
     /**
-     * Returns the fee required by this proposal.
+     * Returns the number of transactions that this proposal will create.
+     *
+     * This is equal to the number of `TransactionSubmitResult`s that will be returned
+     * from `Synchronizer.createProposedTransactions`.
+     *
+     * Proposals always create at least one transaction.
      */
-    fun feeRequired(): Zatoshi {
-        return Zatoshi(inner.feeRequired())
+    fun transactionCount(): Int {
+        return inner.transactionCount()
+    }
+
+    /**
+     * Returns the total fee required by this proposal for its transactions.
+     */
+    fun totalFeeRequired(): Zatoshi {
+        return Zatoshi(inner.totalFeeRequired())
     }
 }